new generic nodes added
foe_bubbles_04_11.awk is a modification of foe_bubbles_09_10.awk. There
are new generic nodes introduced, like the propagator PROP,
distributors DISA, DISFI, DISL, DISFO which have distribution moves
like those of A, FI, L, FO. Also there are nodes MULIN and MULOUT which
are simulating the separation of the molecule into parts which
communicate ones with the others.
The file gen_check_1_foe.awk is a modification of check_1_foe.awk,
where the new nodes are taken care of in the preparation of the json
files. Moreover, borrowing an idea from the space_* scripts, there are
some edges which are deleted in the visualisation, with the purpose to
help the user make sense of what he sees. In this case the edge which
connects to a port in of the MULOUT, and the edge which connects to a
port out of the MULIN are deleted from the json files.
The script gen_viral_foe_bubbles.sh is the same as
main_viral_foe_bubbles.sh, with the only difference that it uses the
new gen_*.awk scripts instead of the older *.awk scripts.