added more quines
the 28_quine.mol is the one which gives the walker, but is not
primitive, one can erase 20 nodes and gets a smaller quine, the
20_quine.mol. The erased part is just a piece of trail, with a bit on
it (or, the walker eats bits).
the 3bit.mol illustrates how can one make a Y combinator like bit
inactive, by trapping it into a quine.
This mechanism is applied for a Turing neuron, by appending at the end
of dendrites some well chosen molecules (identity and a bit,
respectively). The file is tame_turing_neuron.mol. It gives a 22-quine.
At a closer look the walker almost does this for a Y combinator which
generates the trail.
Finally, the 9_quine.mol is one of the smallest quines (can’t be
smaller than 9 nodes, because they need two dist moves and two
compensating beta and fan-in moves). This quine, as the 28-quine, are
introduced in the post
https://chorasimilarity.wordpress.com/2014/09/01/quines-in-chemlambda/