string of 3 20-quines gives quine with the same period
the 20-quine has a period of 4 cycles, which can be checked by running
quiner.sh with wei_*=0. This string of 3 crippled 20-quines is
assembled by replacing  the A-L pattern by its  complement in the
20-quine, then replacing the A-FOE pattern from the complement just
introduced, by its complement. Still the deterministic running shows
that after few steps it stabilises into a quine with 4 cycles period,
just. like the original 20-quine.