commented lines 175-177
These lines shrink all nodes to a radius r=2, after a delay. The
purpose is to make more visible new nodes over old nodes, but since
this may lead to confusion, now these lines are commented.
There is still active the “metabolism” option in the .awk scripts. The
parameter metabo rules the periodic change of colour (to black or back
to original) of old nodes.