a simplification of the previous fibo.mol
Even simplified, it turns out that it works in most of the random runs,
but it does not work properly in the deterministic run! In the
fibo.html demo you see a successful random run.
See also fibo_block_2.mol which works in the deterministic case and in
most of the random runs.